don't cleanup temporary packing directories when using skip_cleanup flag

+++ test/lib/create-next-install.js
@@ -163,7 +163,9 @@ async function createNextInstall({
           })
       }
 
-      await fs.remove(tmpRepoDir)
+      if (!process.env.NEXT_TEST_SKIP_CLEANUP) {
+        await fs.remove(tmpRepoDir)
+      }
       return installDir
     })
 }
